Remove unhelpful prefix code
Before the fix, this code didn't work (as described by the documenting tests).  After the fix, it will never do anything -  can never return a valid value if `Seek` cannot provide one.  The `len` check was also removed, as this constructor is only ever called if `len(prefix)` is non-zero.